+++ previous-versions/accounting/src/sdk/models/operations/getaccounttransaction.ts
@@ -7,12 +7,21 @@ import * as shared from "../shared";
 import { AxiosResponse } from "axios";
 
 export class GetAccountTransactionRequest extends SpeakeasyBase {
+    /**
+     * Unique identifier for an account transaction.
+     */
     @SpeakeasyMetadata({ data: "pathParam, style=simple;explode=false;name=accountTransactionId" })
     accountTransactionId: string;
 
+    /**
+     * Unique identifier for a company.
+     */
     @SpeakeasyMetadata({ data: "pathParam, style=simple;explode=false;name=companyId" })
     companyId: string;
 
+    /**
+     * Unique identifier for a connection.
+     */
     @SpeakeasyMetadata({ data: "pathParam, style=simple;explode=false;name=connectionId" })
     connectionId: string;
 }
@@ -24,6 +33,9 @@ export class GetAccountTransactionResponse extends SpeakeasyBase {
     @SpeakeasyMetadata()
     accountTransaction?: shared.AccountTransaction;
 
+    /**
+     * HTTP response content type for this operation
+     */
     @SpeakeasyMetadata()
     contentType: string;
 
@@ -33,9 +45,15 @@ export class GetAccountTransactionResponse extends SpeakeasyBase {
     @SpeakeasyMetadata()
     errorMessage?: shared.ErrorMessage;
 
+    /**
+     * HTTP response status code for this operation
+     */
     @SpeakeasyMetadata()
     statusCode: number;
 
+    /**
+     * Raw HTTP response; suitable for custom response parsing
+     */
     @SpeakeasyMetadata()
     rawResponse?: AxiosResponse;
 }