+++ previous-versions/accounting/src/sdk/models/operations/listdirectincomeattachments.ts
@@ -7,14 +7,20 @@ import * as shared from "../shared";
 import { AxiosResponse } from "axios";
 
 export class ListDirectIncomeAttachmentsRequest extends SpeakeasyBase {
+    /**
+     * Unique identifier for a company.
+     */
     @SpeakeasyMetadata({ data: "pathParam, style=simple;explode=false;name=companyId" })
     companyId: string;
 
+    /**
+     * Unique identifier for a connection.
+     */
     @SpeakeasyMetadata({ data: "pathParam, style=simple;explode=false;name=connectionId" })
     connectionId: string;
 
     /**
-     * Unique identifier for a direct income
+     * Unique identifier for a direct income.
      */
     @SpeakeasyMetadata({ data: "pathParam, style=simple;explode=false;name=directIncomeId" })
     directIncomeId: string;
@@ -27,6 +33,9 @@ export class ListDirectIncomeAttachmentsResponse extends SpeakeasyBase {
     @SpeakeasyMetadata()
     attachmentsDataset?: shared.AttachmentsDataset;
 
+    /**
+     * HTTP response content type for this operation
+     */
     @SpeakeasyMetadata()
     contentType: string;
 
@@ -36,9 +45,15 @@ export class ListDirectIncomeAttachmentsResponse extends SpeakeasyBase {
     @SpeakeasyMetadata()
     errorMessage?: shared.ErrorMessage;
 
+    /**
+     * HTTP response status code for this operation
+     */
     @SpeakeasyMetadata()
     statusCode: number;
 
+    /**
+     * Raw HTTP response; suitable for custom response parsing
+     */
     @SpeakeasyMetadata()
     rawResponse?: AxiosResponse;
 }