+++ previous-versions/accounting/src/sdk/models/operations/updateinvoice.ts
@@ -10,9 +10,15 @@ export class UpdateInvoiceRequest extends SpeakeasyBase {
     @SpeakeasyMetadata({ data: "request, media_type=application/json" })
     invoice?: shared.Invoice;
 
+    /**
+     * Unique identifier for a company.
+     */
     @SpeakeasyMetadata({ data: "pathParam, style=simple;explode=false;name=companyId" })
     companyId: string;
 
+    /**
+     * Unique identifier for a connection.
+     */
     @SpeakeasyMetadata({ data: "pathParam, style=simple;explode=false;name=connectionId" })
     connectionId: string;
 
@@ -23,16 +29,22 @@ export class UpdateInvoiceRequest extends SpeakeasyBase {
     forceUpdate?: boolean;
 
     /**
-     * Unique identifier for an invoice
+     * Unique identifier for an invoice.
      */
     @SpeakeasyMetadata({ data: "pathParam, style=simple;explode=false;name=invoiceId" })
     invoiceId: string;
 
+    /**
+     * Time limit for the push operation to complete before it is timed out.
+     */
     @SpeakeasyMetadata({ data: "queryParam, style=form;explode=true;name=timeoutInMinutes" })
     timeoutInMinutes?: number;
 }
 
 export class UpdateInvoiceResponse extends SpeakeasyBase {
+    /**
+     * HTTP response content type for this operation
+     */
     @SpeakeasyMetadata()
     contentType: string;
 
@@ -42,9 +54,15 @@ export class UpdateInvoiceResponse extends SpeakeasyBase {
     @SpeakeasyMetadata()
     errorMessage?: shared.ErrorMessage;
 
+    /**
+     * HTTP response status code for this operation
+     */
     @SpeakeasyMetadata()
     statusCode: number;
 
+    /**
+     * Raw HTTP response; suitable for custom response parsing
+     */
     @SpeakeasyMetadata()
     rawResponse?: AxiosResponse;