+++ previous-versions/accounting/src/sdk/models/operations/uploaddirectincomeattachment.ts
@@ -20,17 +20,29 @@ export class UploadDirectIncomeAttachmentRequest extends SpeakeasyBase {
     })
     requestBody?: UploadDirectIncomeAttachmentRequestBody;
 
+    /**
+     * Unique identifier for a company.
+     */
     @SpeakeasyMetadata({ data: "pathParam, style=simple;explode=false;name=companyId" })
     companyId: string;
 
+    /**
+     * Unique identifier for a connection.
+     */
     @SpeakeasyMetadata({ data: "pathParam, style=simple;explode=false;name=connectionId" })
     connectionId: string;
 
+    /**
+     * Unique identifier for a direct income.
+     */
     @SpeakeasyMetadata({ data: "pathParam, style=simple;explode=false;name=directIncomeId" })
     directIncomeId: string;
 }
 
 export class UploadDirectIncomeAttachmentResponse extends SpeakeasyBase {
+    /**
+     * HTTP response content type for this operation
+     */
     @SpeakeasyMetadata()
     contentType: string;
 
@@ -40,9 +52,15 @@ export class UploadDirectIncomeAttachmentResponse extends SpeakeasyBase {
     @SpeakeasyMetadata()
     errorMessage?: shared.ErrorMessage;
 
+    /**
+     * HTTP response status code for this operation
+     */
     @SpeakeasyMetadata()
     statusCode: number;
 
+    /**
+     * Raw HTTP response; suitable for custom response parsing
+     */
     @SpeakeasyMetadata()
     rawResponse?: AxiosResponse;
 }