+++ previous-versions/accounting/src/sdk/models/shared/invoices.ts
@@ -13,10 +13,16 @@ export class Invoices extends SpeakeasyBase {
     @Type(() => Links)
     links: Links;
 
+    /**
+     * Current page number.
+     */
     @SpeakeasyMetadata()
     @Expose({ name: "pageNumber" })
     pageNumber: number;
 
+    /**
+     * Number of items to return in results array.
+     */
     @SpeakeasyMetadata()
     @Expose({ name: "pageSize" })
     pageSize: number;
@@ -26,6 +32,9 @@ export class Invoices extends SpeakeasyBase {
     @Type(() => Invoice)
     results?: Invoice[];
 
+    /**
+     * Total number of items.
+     */
     @SpeakeasyMetadata()
     @Expose({ name: "totalResults" })
     totalResults: number;