+++ previous-versions/accounting/src/sdk/models/shared/item.ts
@@ -8,6 +8,7 @@ import { InvoiceItem } from "./invoiceitem";
 import { ItemStatus } from "./itemstatus";
 import { ItemType } from "./itemtype";
 import { Metadata } from "./metadata";
+import { SupplementalData } from "./supplementaldata";
 import { Expose, Type } from "class-transformer";
 
 /**
@@ -99,6 +100,18 @@ export class Item extends SpeakeasyBase {
     @SpeakeasyMetadata()
     @Expose({ name: "sourceModifiedDate" })
     sourceModifiedDate?: string;
+
+    /**
+     * Supplemental data is additional data you can include in our standard data types.
+     *
+     * @remarks
+     *
+     * It is referenced as a configured dynamic key value pair that is unique to the accounting platform. [Learn more](https://docs.codat.io/using-the-api/supplemental-data/overview) about supplemental data.
+     */
+    @SpeakeasyMetadata()
+    @Expose({ name: "supplementalData" })
+    @Type(() => SupplementalData)
+    supplementalData?: SupplementalData;
 
     /**
      * Type of the item.